refactor(store): move store higher to be shared
