+++ src/containers/safari/inject.js
@@ -7,7 +7,7 @@ import { dispatchInit } from 'containers/safari/dispatch'
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
-import { store } from './store'
+import { store } from 'store'
 //
 
 export function injectDomElements() {
